Changed all doubles to float \n Added option to disable visualization of point clouds for RAM saving

+++ src/structure_refinement/kd_utils.h
@@ -8,8 +8,8 @@ IteratorType split(const IteratorType& begin, const IteratorType& end, const Pre
   IteratorType lower = begin;
   std::reverse_iterator<IteratorType> upper = std::make_reverse_iterator(end); 
   while (lower != upper.base()) {
-    Eigen::Vector3d& v_lower = *lower;
-    Eigen::Vector3d& v_upper = *upper;
+    Eigen::Vector3f& v_lower = *lower;
+    Eigen::Vector3f& v_upper = *upper;
     if (predicate(v_lower)) {
       ++lower;
     }
@@ -22,7 +22,7 @@ IteratorType split(const IteratorType& begin, const IteratorType& end, const Pre
 }
 
 template <typename IteratorType>
-int computeMeanAndCovariance(Eigen::Vector3d& mean, Eigen::Matrix3d& cov,
+int computeMeanAndCovariance(Eigen::Vector3f& mean, Eigen::Matrix3f& cov,
   const IteratorType& begin, const IteratorType& end) {
   // mean computed as 1/(end-start) Sum_k={start..end} x_k
   // cov computed as  (1/(end-start) Sum_k={start..end} x_k* x_k^T ) - mean*mean.transpose();
@@ -30,7 +30,7 @@ int computeMeanAndCovariance(Eigen::Vector3d& mean, Eigen::Matrix3d& cov,
   cov.setZero();
   int k = 0;
   for (IteratorType it = begin; it != end; ++it) {
-    const Eigen::Vector3d& v = *it;
+    const Eigen::Vector3f& v = *it;
     mean += v;
     cov += v * v.transpose();
     ++k;
@@ -38,24 +38,24 @@ int computeMeanAndCovariance(Eigen::Vector3d& mean, Eigen::Matrix3d& cov,
   mean *= (1./k);
   cov *= (1./k);
   cov -= mean * mean.transpose();
-  cov *= double(k)/double(k-1);
+  cov *= float(k)/float(k-1);
   
   return k;
 }
 
 template <typename IteratorType>
-int computeBoundingBox(Eigen::Vector3d& bbox, const Eigen::Vector3d& center,
-  const Eigen::Matrix3d& R, const IteratorType& begin, const IteratorType& end) {
+int computeBoundingBox(Eigen::Vector3f& bbox, const Eigen::Vector3f& center,
+  const Eigen::Matrix3f& R, const IteratorType& begin, const IteratorType& end) {
   bbox.setZero();
   int k = 0;
   
-  Eigen::Vector3d bbox_neg = Eigen::Vector3d::Zero();
-  Eigen::Vector3d bbox_pos = Eigen::Vector3d::Zero();
+  Eigen::Vector3f bbox_neg = Eigen::Vector3f::Zero();
+  Eigen::Vector3f bbox_pos = Eigen::Vector3f::Zero();
   for (IteratorType it = begin; it != end; ++it) {
-    const Eigen::Vector3d v = R*(*it-center);
+    const Eigen::Vector3f v = R*(*it-center);
     for (int i = 0; i < 3; ++i) {
-      bbox_neg(i) = std::min<double>(bbox_neg(i),v(i));
-      bbox_pos(i) = std::max<double>(bbox_pos(i),v(i));
+      bbox_neg(i) = std::min<float>(bbox_neg(i),v(i));
+      bbox_pos(i) = std::max<float>(bbox_pos(i),v(i));
     }
     ++k;
   }
