+++ src/structure_refinement/surfel.cpp
@@ -11,7 +11,7 @@ bool Surfel::putMessage(srrg2_core::BaseSensorMessagePtr msg_) {
   return true;
 }
 
-void Surfel::addObservation(Eigen::Isometry3d& pose, unsigned int poseId, Eigen::Matrix<double, 3, 5>& observation) {
+void Surfel::addObservation(Eigen::Isometry3f& pose, unsigned int poseId, Eigen::Matrix<float, 3, 5>& observation) {
   // Sometimes the same pose and observation might be added twice
   // It hapens when two surfels from cloud i have the same corresponding surfel in cloud j
   odomPoses_.push_back(pose);
@@ -38,7 +38,7 @@ bool Surfel::checkIfPoseExists(unsigned int poseId) {
 
 float Surfel::getLargestRadius() {
   float maxRadius = 0.1;
-  for (Eigen::Matrix<double, 3, 5> observ : observations_) {
+  for (Eigen::Matrix<float, 3, 5> observ : observations_) {
     float tmpRadius = std::sqrt(observ(1, 4) * observ(1, 4) + observ(2, 4) * observ(2, 4));
     if (tmpRadius > maxRadius)
       maxRadius = tmpRadius;
@@ -46,7 +46,7 @@ float Surfel::getLargestRadius() {
   return maxRadius;
 }
 
-void SynthSurfel::addObservation(Eigen::Matrix<double, 3, 2>& observation) {
+void SynthSurfel::addObservation(Eigen::Matrix<float, 3, 2>& observation) {
   // Sometimes the same pose and observation might be added twice
   // It hapens when two surfels from cloud i have the same corresponding surfel in cloud j
   observations_.push_back(observation);
@@ -55,19 +55,19 @@ void SynthSurfel::addObservation(Eigen::Matrix<double, 3, 2>& observation) {
 Eigen::Isometry3f SynthSurfel::getIsometry() {
   Eigen::Isometry3f iso = Eigen::Isometry3f::Identity();
   iso.translation() = this->estPosition_.cast<float>();
-  iso.linear() = matrixBetween2Vectors(Eigen::Vector3d(0, 0, 1), this->estNormal_).cast<float>();
+  iso.linear() = matrixBetween2Vectors(Eigen::Vector3f(0, 0, 1), this->estNormal_).cast<float>();
   return iso;
 }
 
-Eigen::Matrix3d SynthSurfel::matrixBetween2Vectors(Eigen::Vector3d a, Eigen::Vector3d b) {
+Eigen::Matrix3f SynthSurfel::matrixBetween2Vectors(Eigen::Vector3f a, Eigen::Vector3f b) {
   a = a / a.norm();
   b = b / b.norm();
-  Eigen::Vector3d v = a.cross(b);
+  Eigen::Vector3f v = a.cross(b);
   float s = v.norm();
   float c = a.dot(b);
-  Eigen::Matrix3d vx;
+  Eigen::Matrix3f vx;
   vx << 0, -v[2], v[1], v[2], 0, -v[0], -v[1], v[0], 0;
-  Eigen::Matrix3d r = Eigen::Matrix3d::Identity(3, 3);
+  Eigen::Matrix3f r = Eigen::Matrix3f::Identity(3, 3);
   if (s != 0) {
     r = r + vx + vx * vx * ((1 - c) / std::pow(s, 2));
   } else {
